+++ packages/extension-web-ui/src/components/Field/Earning/EarningPoolSelector.tsx
@@ -13,9 +13,9 @@ import { FilterModal } from '@subwallet/extension-web-ui/components/Modal/Filter
 import { SortingModal } from '@subwallet/extension-web-ui/components/Modal/SortingModal';
 import { useFilterModal, useGetPoolTargetList, useYieldPositionDetail } from '@subwallet/extension-web-ui/hooks';
 import { NominationPoolDataType, ThemeProps } from '@subwallet/extension-web-ui/types';
-import { ActivityIndicator, Badge, Button, Icon, InputRef, ModalContext, useExcludeModal } from '@subwallet/react-ui';
+import { Badge, Button, Icon, InputRef, ModalContext, useExcludeModal } from '@subwallet/react-ui';
 import BigN from 'bignumber.js';
-import { Book, CaretLeft, FadersHorizontal, Lightning, SortAscending } from 'phosphor-react';
+import { Book, CaretLeft, FadersHorizontal, SortAscending } from 'phosphor-react';
 import React, { ForwardedRef, forwardRef, SyntheticEvent, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
@@ -56,9 +56,8 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
   const { chain, className = '', defaultValue, disabled,
     from,
     id = 'pool-selector',
-    label, loading, onChange,
+    label, onChange,
     onClickBookButton,
-    placeholder,
     setForceFetchValidator,
     slug, statusHelp,
     value } = props;
@@ -131,9 +130,22 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
             return a.memberCounter - b.memberCounter;
           case SortKey.TOTAL_POOLED:
             return new BigN(b.bondedAmount).minus(a.bondedAmount).toNumber();
-          case SortKey.DEFAULT:
+
           default:
-            return 0;
+            if (sortSelection === SortKey.DEFAULT) {
+              const isSubwalletA = a.name && a.name.includes('SubWallet');
+              const isSubwalletB = b.name && b.name.includes('SubWallet');
+
+              if (isSubwalletA && !isSubwalletB) {
+                return -1;
+              } else if (!isSubwalletA && isSubwalletB) {
+                return 1;
+              }
+
+              return 0;
+            } else {
+              return 0;
+            }
         }
       });
   }, [items, selectedFilters, sortSelection]);
@@ -189,10 +201,21 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
   }, [chain, items.length, setForceFetchValidator, t]);
 
   const renderSelected = useCallback((item: NominationPoolDataType) => {
+    const isCheckRecommend = item.name?.includes('SubWallet') || false;
+
     return (
       <div className={'__selected-item'}>
         <div className={'__selected-item-name common-text'}>
-          {item.name || `Pool #${item.id}`}
+          {isCheckRecommend
+            ? (
+              <>
+                {item.name}
+                <div className={'__title-suffix'}>&nbsp;(Recommended)</div>
+              </>
+            )
+            : (
+              item.name || `Pool #${item.id}`
+            )}
         </div>
       </div>
     );
@@ -213,17 +236,6 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
   const onCloseDetail = useCallback(() => {
     inactiveModal(EarningPoolDetailModalId);
   }, [inactiveModal]);
-
-  const onClickLightningButton = useCallback((e: SyntheticEvent) => {
-    e.stopPropagation();
-    const poolId = defaultSelectPool;
-
-    if (poolId !== undefined) {
-      onChange?.({ target: { value: `${poolId}` } });
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [slug]);
 
   useEffect(() => {
     const defaultSelectedPool = defaultValue || nominationPoolValueList[0] || `${defaultSelectPool || ''}`;
@@ -265,10 +277,10 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
         itemKey={'idStr'}
         items={resultList}
         label={label}
-        loading={loading}
+        loading={false}
         onClickActionBtn={onClickActionBtn}
         onSelect={_onSelectItem}
-        placeholder={placeholder || t('Select pool')}
+        placeholder={t('Select pool')}
         prefix={(
           <Avatar
             size={20}
@@ -295,45 +307,23 @@ const Component = (props: Props, ref: ForwardedRef<InputRef>) => {
         selected={value || ''}
         showActionBtn
         statusHelp={statusHelp}
-        suffix={loading
-          ? (
-            <div>
-              <ActivityIndicator size={20} />
-            </div>
-          )
-          : (
-            <div className='select-pool-suffix'>
-              <Button
-                disabled={isDisabled}
-                icon={(
-                  <Icon
-                    phosphorIcon={Book}
-                    size='sm'
-                  />
-                )}
-                onClick={onClickBookButton}
-                size='xs'
-                type='ghost'
-              />
-              {
-                !!defaultSelectPool && (
-                  <Button
-                    disabled={isDisabled}
-                    icon={(
-                      <Icon
-                        phosphorIcon={Lightning}
-                        size='sm'
-                      />
-                    )}
-                    onClick={onClickLightningButton}
-                    size='xs'
-                    type='ghost'
-                  />
-                )
-              }
-            </div>
-          )}
-        title={label || placeholder || t('Select pool')}
+        suffix={(
+          <div className='select-pool-suffix'>
+            <Button
+              disabled={isDisabled}
+              icon={(
+                <Icon
+                  phosphorIcon={Book}
+                  size='sm'
+                />
+              )}
+              onClick={onClickBookButton}
+              size='xs'
+              type='ghost'
+            />
+          </div>
+        )}
+        title={t('Select pool')}
       />
 
       <FilterModal
@@ -385,6 +375,16 @@ const EarningPoolSelector = styled(forwardRef(Component))<Props>(({ theme: { tok
         paddingTop: 0,
         paddingBottom: token.paddingXXS
       }
+    },
+    '.__title-suffix': {
+      fontSize: token.fontSizeSM,
+      fontWeight: token.bodyFontWeight,
+      lineHeight: token.lineHeightSM,
+      color: token.colorTextTertiary
+    },
+    '.__selected-item-name.common-text': {
+      display: 'flex',
+      alignItems: 'baseline'
     },
 
     '.ant-select-modal-input-wrapper': {